LeoMine/requirements.txt
appdirs==1.4.4
attrs==19.3.0
beautifulsoup4==4.9.1
black==19.10b0
blis==0.4.1
boto==2.49.0
boto3==1.14.23
botocore==1.17.23
bs4==0.0.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
cymem==2.0.3
dnspython==2.0.0
docutils==0.15.2
en-core-web-sm==2.3.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
Flask-Cors==3.0.8
flask-ngrok==0.0.25
gensim==3.8.3
idna==2.10
importlib-metadata==1.7.0
isort==5.1.4
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
lxml==4.5.2
MarkupSafe==1.1.1
murmurhash==1.0.2
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.0
pandas==1.0.5
pathspec==0.8.0
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
pyenchant==3.1.1
pymongo==3.10.1
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-file==1.5.1
s3transfer==0.3.3
scipy==1.5.1
six==1.15.0
smart-open==2.1.0
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.2
thinc==7.4.1
tinysegmenter==0.3
tldextract==2.2.2
toml==0.10.1
tqdm==4.48.0
typed-ast==1.4.1
urllib3==1.25.9
wasabi==0.7.1
Werkzeug==1.0.1
zipp==3.1.0
